+++ io_import_rbsp/rbsp/apex_legends/__init__.py
@@ -1,4 +1,6 @@
 __all__ = ["entities", "geometry", "materials", "props"]
+from typing import List
+
 import bmesh
 import bpy
 
@@ -6,7 +8,7 @@ from ..titanfall import entities, props
 from . import materials
 
 
-def geometry(bsp, master_collection, materials):
+def geometry(bsp, master_collection: bpy.types.Collection, materials: List[bpy.types.Material]):
     geometry_collection = bpy.data.collections.new("geometry")
     master_collection.children.link(geometry_collection)
     # load_model