+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog
 
+## v1.2.0_b2.93
+
+### Added
+ * Entity naming resolution order is now `targetname -> editorclass -> classname`
+
+### TODOS
+ * colour palette for trigger brushes
+   - 1 colour for each classname / editorclass
+
 ## v1.1.0_b2.93 (1st October 2021)
 
 ### Added
  * Geometry for `trigger_*` brush entities is now generated on import
+
 
 ## v1.0.0_b2.93 (28th September 2021)
 Initial Release